# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sinema API",
        "description": "Film, tür, kişi ve oyuncu yönetimi için RESTful API",
        "contact": {
            "name": "API Geliştirici",
            "email": "developer@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/oyuncular": {
            "get": {
                "tags": [
                    "Oyuncular"
                ],
                "summary": "Tüm oyuncuları listele",
                "description": "Veritabanındaki tüm oyuncu-film ilişkilerini getirir",
                "operationId": "d446ad564d3ae4692ae4460b84ebbf9b",
                "responses": {
                    "200": {
                        "description": "Başarılı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Oyuncu"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Oyuncular"
                ],
                "summary": "Yeni oyuncu ekle",
                "description": "Veritabanına yeni bir oyuncu-film ilişkisi ekler",
                "operationId": "ee7036ab9f47f549c725645875874555",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "film_id",
                                    "kisi_id"
                                ],
                                "properties": {
                                    "film_id": {
                                        "description": "Film ID'si",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "kisi_id": {
                                        "description": "Kişi ID'si",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Oyuncu başarıyla eklendi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Oyuncu"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Doğrulama hatası"
                    }
                }
            }
        },
        "/oyuncular/{id}": {
            "get": {
                "tags": [
                    "Oyuncular"
                ],
                "summary": "Belirli oyuncuyu getir",
                "description": "ID'ye göre belirli bir oyuncu-film ilişkisini getirir",
                "operationId": "bffa9460dd1d4dc05721622785f07ced",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Oyuncu ID'si",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Başarılı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Oyuncu"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Oyuncu bulunamadı"
                    }
                }
            },
            "put": {
                "tags": [
                    "Oyuncular"
                ],
                "summary": "Oyuncu güncelle",
                "description": "Belirli bir oyuncu-film ilişkisini günceller",
                "operationId": "36d6ca873c116bf5be80cc9938e3ca2e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Oyuncu ID'si",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "film_id",
                                    "kisi_id"
                                ],
                                "properties": {
                                    "film_id": {
                                        "description": "Film ID'si",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "kisi_id": {
                                        "description": "Kişi ID'si",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Oyuncu başarıyla güncellendi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Oyuncu"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Oyuncu bulunamadı"
                    },
                    "422": {
                        "description": "Doğrulama hatası"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Oyuncular"
                ],
                "summary": "Oyuncu sil",
                "description": "Belirli bir oyuncu-film ilişkisini veritabanından siler",
                "operationId": "865d5e6282dff5153a6485c3bfadd52f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Oyuncu ID'si",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Oyuncu başarıyla silindi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Oyuncu başarıyla silindi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Oyuncu bulunamadı"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Film": {
                "title": "Film",
                "description": "Film modeli",
                "properties": {
                    "id": {
                        "description": "Benzersiz ID",
                        "type": "integer",
                        "example": 1
                    },
                    "adi": {
                        "description": "Film adı",
                        "type": "string",
                        "example": "The Matrix"
                    },
                    "konusu": {
                        "description": "Film konusu",
                        "type": "string",
                        "example": "Bir bilgisayar programcısı..."
                    },
                    "imdb_puani": {
                        "description": "IMDB puanı",
                        "type": "number",
                        "format": "float",
                        "example": 8.7
                    },
                    "tur_id": {
                        "description": "Tür ID'si",
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "tur": {
                        "$ref": "#/components/schemas/Tur"
                    },
                    "oyuncular": {
                        "description": "Film oyuncuları",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Oyuncu"
                        }
                    }
                },
                "type": "object"
            },
            "Kisi": {
                "title": "Kişi",
                "description": "Kişi modeli (oyuncu, yönetmen vs.)",
                "properties": {
                    "id": {
                        "description": "Benzersiz ID",
                        "type": "integer",
                        "example": 1
                    },
                    "adi": {
                        "description": "Kişi adı",
                        "type": "string",
                        "example": "Keanu Reeves"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Oyuncu": {
                "title": "Oyuncu",
                "description": "Film-kişi ilişki modeli",
                "properties": {
                    "id": {
                        "description": "Benzersiz ID",
                        "type": "integer",
                        "example": 1
                    },
                    "film_id": {
                        "description": "Film ID'si",
                        "type": "integer",
                        "example": 1
                    },
                    "kisi_id": {
                        "description": "Kişi ID'si",
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "film": {
                        "$ref": "#/components/schemas/Film"
                    },
                    "kisi": {
                        "$ref": "#/components/schemas/Kisi"
                    }
                },
                "type": "object"
            },
            "Tur": {
                "title": "Tür",
                "description": "Film türü modeli",
                "properties": {
                    "id": {
                        "description": "Benzersiz ID",
                        "type": "integer",
                        "example": 1
                    },
                    "adi": {
                        "description": "Tür adı",
                        "type": "string",
                        "example": "Aksiyon"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-26T20:27:34.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-26T20:27:34.000000Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}